backend/src/requirements.txt
# This file lists the project's dependencies.
psycopg2-binary
requests
openai
groq
scikit-learn
dotenv